+++ datawolf-executor-hpc/src/main/java/edu/illinois/ncsa/datawolf/executor/hpc/HPCExecutor.java
@@ -205,6 +205,10 @@ public class HPCExecutor extends RemoteExecutor {
                             String optionId = option.getOptionId();
                             Map<String, String> inputs = step.getInputs();
                             String key = inputs.get(optionId);
+                            if (execution.getDataset(key).isEmpty()) {
+                                // No dataset has been set, must be an optional dataset so skip it
+                                break;
+                            }
                             Dataset ds = datasetDao.findOne(execution.getDataset(key));
                             if (ds == null) {
                                 throw (new AbortException("Dataset is missing."));