+++ datawolf-executor-kubernetes/src/main/java/edu/illinois/ncsa/datawolf/executor/kubernetes/KubernetesExecutor.java
@@ -142,6 +142,11 @@ public class KubernetesExecutor extends RemoteExecutor {
 
                     if (option.getInputOutput() != InputOutput.OUTPUT) {
                         String key = step.getInputs().get(option.getOptionId());
+                        if (execution.getDataset(key).isEmpty()) {
+                            // No dataset has been set, must be an optional dataset so skip it
+                            break;
+                        }
+
                         Dataset ds = datasetDao.findOne(execution.getDataset(key));// option.getOptionId()));
                         if (ds == null) {
                             throw (new AbortException("Dataset is missing."));
@@ -470,11 +475,19 @@ public class KubernetesExecutor extends RemoteExecutor {
                                     }
                                 });
 
-                                for (File file : files) {
-                                    logger.debug("adding files to a dataset: " + file);
-                                    FileInputStream fis = new FileInputStream(file);
-                                    fileStorage.storeFile(file.getName(), fis, execution.getCreator(), ds);
-                                    fis.close();
+                                if (files != null) {
+                                    if (files.length == 0 && (!step.getTool().getOutput(entry.getKey()).isAllowNull())) {
+                                        // Required output was not found, fail the step
+                                        logger.error("Could not find required output files, failing the workflow step.");
+                                        throw new FailedException("Required output files are missing.");
+                                    }
+
+                                    for (File file : files) {
+                                        logger.debug("adding files to a dataset: " + file);
+                                        FileInputStream fis = new FileInputStream(file);
+                                        fileStorage.storeFile(file.getName(), fis, execution.getCreator(), ds);
+                                        fis.close();
+                                    }
                                 }
 
                             } else {