+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Build failure from update to maven assembly plugin configuration element [#44](https://github.com/ncsa/datawolf/issues/44)
 
+### Changed
+- WorkflowToolData to allow optional datasets [#39](https://github.com/ncsa/datawolf/issues/39)
+
 
 ## [4.7.0] - 2024-10-29