+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - WorkflowToolData to allow optional datasets [#39](https://github.com/ncsa/datawolf/issues/39)
+- IncoreFileStorage to use the filename as the dataset name [#47](https://github.com/ncsa/datawolf/issues/47)
+
+### Fixed
+- Build failure from update to maven assembly plugin configuration element [#44](https://github.com/ncsa/datawolf/issues/44)
 
 
 ## [4.7.0] - 2024-10-29